+++ arrangehttp/server_test.go
@@ -11,7 +11,6 @@ import (
 	"net"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"strconv"
 	"strings"
 	"testing"
@@ -22,6 +21,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/xmidt-org/arrange"
+	"github.com/xmidt-org/arrange/arrangetls"
 	"go.uber.org/fx"
 	"go.uber.org/fx/fxtest"
 )
@@ -85,8 +85,6 @@ func testServerConfigTLS(t *testing.T) {
 		assert  = assert.New(t)
 		require = require.New(t)
 
-		certificateFile, keyFile = createServerFiles(t)
-
 		serverConfig = ServerConfig{
 			Address:           ":0",
 			ReadTimeout:       72 * time.Second,
@@ -95,11 +93,11 @@ func testServerConfigTLS(t *testing.T) {
 			IdleTimeout:       9234 * time.Minute,
 			MaxHeaderBytes:    3642,
 			KeepAlive:         3 * time.Minute,
-			TLS: &TLS{
-				Certificates: ExternalCertificates{
+			TLS: &arrangetls.Config{
+				Certificates: arrangetls.ExternalCertificates{
 					{
-						CertificateFile: certificateFile,
-						KeyFile:         keyFile,
+						CertificateFile: CertificateFile,
+						KeyFile:         KeyFile,
 					},
 				},
 			},
@@ -108,9 +106,6 @@ func testServerConfigTLS(t *testing.T) {
 		router  = mux.NewRouter()
 		address = make(chan net.Addr, 1)
 	)
-
-	defer os.Remove(certificateFile)
-	defer os.Remove(keyFile)
 
 	server, listen, err := serverConfig.NewServer()
 	require.NoError(err)